fix(autoscaler): pass smoothed busy/idle as floats to PoolSnapshot
Smoothed busy was rounded to int before reaching the strategy, so a
worker busy ~50% of the time toggled between 0 and 1, which
ceil(busy/target) then amplified into target=1↔2 oscillation regardless
of cooldowns. Keep the EWMA value as a float through to the strategy.

+++ src/Autoscaler/PoolSnapshot.php
@@ -9,8 +9,8 @@ final readonly class PoolSnapshot
     public function __construct(
         public string $transport,
         public int $currentWorkers,
-        public int $busyWorkers,
-        public int $idleWorkers,
+        public float $busyWorkers,
+        public float $idleWorkers,
         public float $throughputPerSecond,
         public ?int $queueDepth,
         public int $min,
